Fix goroutine/socket leak in multiplexer
Fixes gravitational/teleport-private#78
LAT-APP21-3

Change the multiplexer from opt-out to opt-in for protocol listeners.

The multiplexer previously created new listeners for each protocol (SSH,
TLS, DB) and its Config contained opt-out configurations (DisableTLS,
DisableSSH, DisablePostgres). When callers didn't explicitly disable a
protocol, new connections for that protocol would leak a goroutine and
fail to close connections. This exposed a 3-line DoS whereby multiple
connections could be passed to the multiplexer for a protocol that was
not being serviced, eventually resulting in file descriptor limits being
hit which then prevented the teleport process from operating (see
example in issue).

Changing to opt-in means new protocols can be added to the multiplexer
without requiring all existing callers to be changed (to opt-out of the
new protocol). Forgetting to opt-out would expose the new code to DoS
without compile-time, test, or operational notification.

Rather than rely on opt-out flags in the config, this change creates
listeners only when explicitly requested by callers. The existing getter
methods on multiplexer were changed to create listeners when called. And
multiplexer protocol detection now closes connections when a listener
hasn't been created. This also allowed for the protocol detection
routine to be simplified.